Pass jdom Element a bit further from SwordContentFacade
It can be useful when inspecting / manipulating xml
