+++ build.gradle
@@ -4,6 +4,8 @@ buildscript {
     ext.kotlin_version = '1.4.32'
     ext.room_version = '2.3.0'
     ext.jsword_version = '2.3.32'
+    ext.jdom_version = '2.0.6'
+    ext.commons_text_version = '1.9'
     ext.kotlinx_serialization_version = '1.1.0'
 
     repositories {
