+++ db/build.gradle
@@ -31,6 +31,8 @@ dependencies {
     implementation "androidx.room:room-ktx:$room_version"
     implementation "org.jetbrains.kotlinx:kotlinx-serialization-json:$kotlinx_serialization_version"
     implementation "com.github.AndBible:jsword:$jsword_version"
+    implementation "org.jdom:jdom2:$jdom_version"
+    implementation "org.apache.commons:commons-text:$commons_text_version"
 
     kapt "androidx.room:room-compiler:$room_version"
 }
